make layout testing code public
